+++ middlewares/mask-output.mjs
@@ -9,9 +9,9 @@ export function cleanObject(e) {
 
   let item = e;
   if (e.toCleanObject) {
-    item = e.toCleanObject();
+    item = e.toCleanObject({ virtuals: true });
   } else if (e.toObject) {
-    item = e.toObject();
+    item = e.toObject({ virtuals: true });
   }
 
   if (!isPlainObject(item) && !Array.isArray(item)) {
@@ -22,7 +22,7 @@ export function cleanObject(e) {
     if (['__v'].includes(key)) {
       return;
     }
-    acc[key.replace ? key.replace('_', '') : key] = cleanObject(val);
+    acc[key.replace ? key.replace(/^_/g, '') : key] = cleanObject(val);
   });
 }