+++ models/rules.mjs
@@ -1,47 +1,48 @@
 import lGet from 'lodash.get';
+import nanoid from 'nanoid';
 // It's not really a cycle import, we're not importing the same part of the tree
 // eslint-disable-next-line import/no-cycle
 import { CAN_LIST_ALL_CAMPUSES } from './rights';
+import { getPrefetchedRide } from '../helpers/prefetch-ride';
+
+const ruleGenerator = (rule = () => true) => (id) => ({ id: Symbol(id || nanoid()), rule });
+/**
+ * @return {symbol}
+ */
+export const stdRule = ruleGenerator();
 
 /**
  * @return {symbol}
  */
-export const stdRule = (id) => ({
-  id: Symbol(id),
-  rule: () => true,
+export const campusRule = ruleGenerator(({ campuses = [] }, ctx) => {
+  if (ctx.may(CAN_LIST_ALL_CAMPUSES)) {
+    return true;
+  }
+  const campus = lGet(ctx, 'params.campus_id', lGet(ctx, 'query.filters.campus', null));
+  return campus && !!campuses.find((c) => c._id === campus);
 });
 
 /**
  * @return {symbol}
  */
-export const campusRule = (id) => ({
-  id: Symbol(id),
-  rule: ({ campuses = [] }, ctx) => {
-    if (ctx.may(CAN_LIST_ALL_CAMPUSES)) {
-      return true;
-    }
-    const campus = lGet(ctx, 'params.campus_id', lGet(ctx, 'query.filters.campus', null));
-    return campus && !!campuses.find((c) => c._id === campus);
-  },
+export const selfEditingUserRule = ruleGenerator((_, ctx) => {
+  const userParam = lGet(ctx, 'params.user_id', null);
+  const loggedUser = lGet(ctx, 'state.user.id', null);
+  return userParam && loggedUser && userParam === loggedUser;
 });
 
-/**
- * @return {symbol}
- */
-export const selfEditingUserRule = (id) => ({
-  id: Symbol(id),
-  rule: (_, ctx) => {
-    const userParam = lGet(ctx, 'params.user_id', null);
-    const loggedUser = lGet(ctx, 'state.user.id', null);
-    return userParam && loggedUser && userParam === loggedUser;
-  },
-});
+export const roleEditingRule = ruleGenerator((
+  { campuses = [] },
+  ctx,
+  { id: campusId = null } = {},
+) => campusId && !!campuses.find((c) => c._id === campusId));
 
-export const roleEditingRule = (id) => ({
-  id: Symbol(id),
-  rule: (
-    { campuses = [] },
-    ctx,
-    { id: campusId = null } = {},
-  ) => campusId && !!campuses.find((c) => c._id === campusId),
+export const tokenRideRule = ruleGenerator((_, ctx) => {
+  const ride = getPrefetchedRide(ctx, ctx.params.id);
+  return ride.token === ctx.query.token;
+});
+export const ownedRideRule = ruleGenerator((_, ctx) => {
+  const { user } = ctx.state;
+  const ride = getPrefetchedRide(ctx, ctx.params.id);
+  return ride.owner.id.toString() === user.id;
 });