+++ services/config.mjs
@@ -5,6 +5,7 @@ nconf.argv()
   .env({
     separator: '__',
     lowerCase: true,
+    parseValues: true,
   })
   .file({
     file: path.join(path.dirname(new URL(import.meta.url).pathname), '..', 'config.json'),